+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0-RC3] - 2019-11.24
+### Fixed
+- Fixed a bug that needed settings to be saved once, before using the plugin
+- Cancelling background image selection no longer leads to cancelling the metadata download
+
 ## [1.0-RC2] - 2019-11.24
 ### Added
 - Changelog
@@ -17,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/Mithnar/PlayniteVndb/compare/1.0-RC2...HEAD
+[Unreleased]: https://github.com/Mithnar/PlayniteVndb/compare/1.0-RC3...HEAD
+[1.0-RC3]: https://github.com/Mithnar/PlayniteVndb/compare/1.0-RC2...1.0-RC3
 [1.0-RC2]: https://github.com/Mithnar/PlayniteVndb/compare/1.0-RC1...1.0-RC2
 [1.0-RC1]: https://github.com/Mithnar/PlayniteVndb/releases/tag/1.0-RC1